Fix calcPr to avoid false save file prompt
This commit changes the calcId from 124519 to 125725 and removes the
'fullCalcOnLoad' attribute to avoid save file propmt when user exits
with no changes. This issue was reported as issue #100. The new calcId
is extracted from Excel file generated by Excel 2007 (12.0.6683.5002)
SP3 MSO (12.0.6683.5000). This fix may not work with Excel versions
later than 2007.