Set the zipfile member timestamps to the workbook timestamp
This is particularly useful when the "created" workbook property is
manually set, as it ensures that a message digest of the generated .xlsx
remains stable across equivalent runs.

When the "created" workbook property is automatically set to the current
time, this shouldn't have any ill-effect as the times were previously
taken from the just-written temporary files.

Time::Local is core since the first release of Perl 5.